Splits the algorithm from the cli
* swagger-min.js is now the main algorithm
* real unit tests have been added
* code coverage requirements are now in place in order to build
