--- test-results.xml
@@ -1,4 +1,0 @@
-<testsuite name="Mocha Tests" tests="2" failures="0" errors="0" skipped="0" timestamp="Mon, 11 Feb 2019 16:09:41 GMT" time="0.011">
-<testcase classname="fake test" name="is a fake test" time="0"/>
-<testcase classname="fake test" name="says hello" time="0"/>
-</testsuite>
